+++ src/game.js
@@ -2,29 +2,39 @@ import DOMRenderer from "./dom-renderer";
 import NullRenderer from "./null-renderer";
 import BoardState from "./board-state";
 import Ruleset from "./ruleset";
+import Scorer from "./scorer";
 
 const VALID_GAME_OPTIONS = [
   "boardSize",
   "scoring",
   "handicapStones",
-  "koRule"
+  "koRule",
+  "_hooks"
 ];
 
 const Game = function(boardElement) {
   this._defaultBoardSize = 19;
   this.boardSize = null;
   this._moves = [];
-  this.renderer = boardElement ? new DOMRenderer(this, boardElement) : new NullRenderer();
   this.callbacks = {
     postRender: function() {}
   };
-  this._deadPoints = [];
+  this._boardElement = boardElement;
   this._defaultScoring = "territory";
   this._defaultKoRule = "simple";
+  this._deadPoints = [];
 };
 
 Game.prototype = {
   _configureOptions: function({ boardSize = this._defaultBoardSize, handicapStones = 0, scoring = this._defaultScoring, koRule = this._defaultKoRule } = {}) {
+    if (typeof boardSize !== "number") {
+        throw new Error("Board size must be a number, but was: " + (typeof boardSize));
+      }
+
+    if (typeof handicapStones !== "number") {
+      throw new Error("Handicap stones must be a number, but was: " + (typeof boardSize));
+    }
+
     if (handicapStones > 0 && boardSize !== 9 && boardSize !== 13 && boardSize !== 19) {
       throw new Error("Handicap stones not supported on sizes other than 9x9, 13x13 and 19x19");
     }
@@ -33,15 +43,26 @@ Game.prototype = {
       throw new Error("Only 2 to 9 handicap stones are supported");
     }
 
+    if (boardSize > 19) {
+      throw new Error("cannot generate a board size greater than 19");
+    }
+
     this.boardSize = boardSize;
     this.handicapStones = handicapStones;
-    this.ruleset = new Ruleset({
-      "scoring": scoring,
-      "koRule": koRule
+    this._scorer = new Scorer({
+      scoreBy: scoring
     });
+
+    this._whiteMustPassLast = this._scorer.usingPassStones();
+
+    this._ruleset = new Ruleset({
+      koRule: koRule
+    });
+
+    this._initialState = BoardState._initialFor(boardSize, handicapStones);
   },
 
-  setup: function(options) {
+  setup: function(options = {}) {
     for (let key in options) {
       if (options.hasOwnProperty(key) && VALID_GAME_OPTIONS.indexOf(key) < 0) {
         throw new Error("Unrecognized game option: " + key);
@@ -50,53 +71,59 @@ Game.prototype = {
 
     this._configureOptions(options);
 
-    if (this.boardSize > 19) {
-      throw new Error("cannot generate a board size greater than 19");
+    if (this._boardElement) {
+      const defaultRendererHooks = {
+        handleClick: (y, x) => {
+          if (this.isOver()) {
+            this.toggleDeadAt(y, x);
+          } else {
+            this.playAt(y, x);
+          }
+        },
+
+        hoverValue: (y, x) => {
+          if (!this.isOver() && !this.isIllegalAt(y, x)) {
+            return this.currentPlayer();
+          }
+        },
+
+        gameIsOver: () => {
+          return this.isOver();
+        }
+      };
+
+      this.renderer = new DOMRenderer(this._boardElement, options["hooks"] || defaultRendererHooks);
+    } else {
+      this.renderer = new NullRenderer();
     }
 
-    this.renderer.setup();
     this.render();
   },
 
   intersectionAt: function(y, x) {
-    return this.boardState().intersectionAt(y, x);
+    return this.currentState().intersectionAt(y, x);
   },
 
   intersections: function() {
-    return this.boardState().intersections;
+    return this.currentState().intersections;
   },
 
-  yCoordinateFor: function(y) {
-    return this.boardSize - y;
-  },
-
-  xCoordinateFor: function(x) {
-    const letters = ["A", "B", "C", "D", "E", "F", "G", "H", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T"];
-
-    return letters[x];
+  deadStones: function() {
+    return this._deadPoints;
   },
 
   coordinatesFor: function(y, x) {
-    return this.xCoordinateFor(x) + this.yCoordinateFor(y);
+    return this.currentState().xCoordinateFor(x) + this.currentState().yCoordinateFor(y);
   },
 
   currentPlayer: function() {
-    return this.boardState()._nextColor();
-  },
+    const lastMoveColor = this.currentState().color;
 
-  playAt: function(y, x) {
-    if (this.isIllegalAt(y, x)) {
-      return false;
+    if (lastMoveColor === "black") {
+      return "white";
+    } else {
+      return "black";
     }
-
-    this._moves.push(this.boardState().playAt(y, x));
-    this.render();
-
-    return true;
-  },
-
-  boardState: function() {
-    return this._moves[this._moves.length - 1] || BoardState._initialFor(this.boardSize, this.handicapStones);
   },
 
   isWhitePlaying: function() {
@@ -107,57 +134,76 @@ Game.prototype = {
     return this.currentPlayer() === "black";
   },
 
-  inAtari: function(y, x) {
-    return this.boardState().inAtari(y, x);
+  score: function() {
+    return this._scorer.score(this);
   },
 
-  wouldBeSuicide: function(y, x) {
-    return this.boardState().wouldBeSuicide(y, x);
+  currentState: function() {
+    return this._moves[this._moves.length - 1] || this._initialState;
+  },
+
+  playAt: function(y, x) {
+    if (this.isIllegalAt(y, x)) {
+      return false;
+    }
+
+    const newState = this.currentState().playAt(y, x, this.currentPlayer());
+    this._moves.push(newState);
+
+    this.render();
+
+    return true;
   },
 
   pass: function() {
-    if (!this.isOver()) {
-      this._moves.push(this.boardState().playPass());
-      this.render();
+    if (this.isOver()) {
+      return false;
     }
+
+    const newState = this.currentState().playPass(this.currentPlayer());
+    this._moves.push(newState);
+
+    this.render();
+
+    return true;
   },
 
   isOver: function() {
-    return this.ruleset.isOver(this);
+    if (this._moves.length < 2) {
+      return false;
+    }
+
+    if (this._whiteMustPassLast) {
+      const finalMove = this._moves[this._moves.length - 1];
+      const previousMove = this._moves[this._moves.length - 2];
+
+      return finalMove.pass && previousMove.pass && finalMove.color === "white";
+    } else {
+      const finalMove = this._moves[this._moves.length - 1];
+      const previousMove = this._moves[this._moves.length - 2];
+
+      return finalMove.pass && previousMove.pass;
+    }
   },
 
   toggleDeadAt: function(y, x) {
-    const alreadyDead = this.isDeadAt(y, x);
+    const alreadyDead = this._isDeadAt(y, x);
 
-    this.groupAt(y, x).forEach(intersection => {
+    this.currentState().groupAt(y, x).forEach(intersection => {
       if (alreadyDead) {
         this._deadPoints = this._deadPoints.filter(dead => !(dead.y === intersection.y && dead.x === intersection.x));
       } else {
-        this._deadPoints.push({ y: intersection.y, x: intersection.x });
+        this._deadPoints.push({ y: intersection.y, x: intersection.x, color: intersection.value });
       }
     });
 
     this.render();
+
+    return true;
   },
 
-  isDeadAt: function(y, x) {
+  _isDeadAt: function(y, x) {
     return this._deadPoints.some(dead => dead.y === y && dead.x === x);
-  },
-
-  score: function() {
-    return this.ruleset.score(this);
-  },
-
-  libertiesAt: function(y, x) {
-    return this.boardState().libertiesAt(y, x);
-  },
-
-  groupAt: function(y, x) {
-    return this.boardState().groupAt(y, x);
-  },
-
-  neighborsFor: function(y, x) {
-    return this.boardState().neighborsFor(y, x);
   },
 
   isIllegalAt: function(y, x) {
@@ -165,33 +211,36 @@ Game.prototype = {
       return false;
     }
 
-    return this.ruleset.isIllegal(y, x, this);
-  },
-
-  render: function() {
-    if (!this.isOver()) {
-      this.removeScoringState();
-    }
-
-    this.renderer.render();
-    this.callbacks.postRender(this);
-  },
-
-  removeScoringState: function() {
-    this._deadPoints = [];
+    return this._ruleset.isIllegal(y, x, this);
   },
 
   territory: function() {
     if (!this.isOver()) {
-      return;
+      return {
+        black: [],
+        white: []
+      };
     }
 
-    return this.ruleset.territory(this);
+    return this._scorer.territory(this);
   },
 
   undo: function() {
     this._moves.pop();
     this.render();
+  },
+
+  render: function() {
+    if (!this.isOver()) {
+      this._deadPoints = [];
+    }
+
+    this.renderer.render(this.currentState(), {
+      territory: this.territory(),
+      deadStones: this.deadStones()
+    });
+
+    this.callbacks.postRender(this);
   }
 };