+++ src/null-renderer.js
@@ -1,4 +1,5 @@
 export default function NullRenderer() {
   this.setup = function() {};
   this.render = function() {};
+  this.renderTerritory = function() {};
 }